Reworked the whole character creator window.
Made a separate window for material creation.
Made the whole user process for configuring and creating an entry a step by step configuration and validation, and then pressing a button creates and sets everything.
Made a configuration window to store default data and a reference to the csv, so it only needs to be set once.